refactor route and added types
